+++ workflow/index.d.ts
@@ -1,4 +1,0 @@
-declare type TemplateFromFile = { readonly type: "from-file"; readonly path: string; };
-declare type TplName = "calculate-pfconv-params" | "export-clones" | "get-preset" | "list-presets" | "mixcr-analyze" | "prerun" | "process" | "test.columns.test" | "main";
-declare const Templates: Record<TplName, TemplateFromFile>;
-export { Templates };