+++ ui/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @milaboratory/milaboratories.mixcr-clonotyping.ui
+# @platforma-open/milaboratories.mixcr-clonotyping.ui
 
 ## 1.0.1
 
@@ -6,4 +6,4 @@
 
 - 3a4df33: Initial release
 - Updated dependencies [3a4df33]
-  - @milaboratory/milaboratories.mixcr-clonotyping.model@1.0.1
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.0.1